+++ docker/backend/Dockerfile
@@ -13,7 +13,7 @@ WORKDIR /app
 # Copy the directories
 COPY ./service ./service
 COPY ./opencopilot ./opencopilot
-COPY ./opencopilot/configs/env-postgres.py ./opencopilot/configs/env.py
+COPY ./service/configs/env-postgres.py ./opencopilot/configs/env.py
 COPY ./start.sh ./
 
 # Install the Python dependencies
