--- service/configs.py
@@ -1,6 +1,0 @@
-# DATABASE CONFIGURATIONS
-db_host = "database"
-db_name = "ecommerce"
-db_user = "root"
-db_password = "root"
-db_port = "5432"
